+++ frontend/src/app/metadataModule/metadata-dataset.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { TranslateService } from "@ngx-translate/core";
 import { MatDialog } from "@angular/material";
+import { BehaviorSubject } from "rxjs";
 import { tap, map } from "rxjs/operators";
 
 import { DataFormService } from '@geonature_common/form/data-form.service';
+import { ModuleService } from '@geonature/services/module.service';
 import { ConfirmationDialog } from "@geonature_common/others/modal-confirmation/confirmation.dialog";
 import { MetadataService } from './services/metadata.service';
 import { MetadataDataService } from './services/metadata-data.service';
@@ -18,18 +20,39 @@ export class MetadataDatasetComponent implements OnInit {
 
   @Input() dataset: any;
   stateChangeSaving: boolean = false;
+
+  private _moduleImportIsAuthorized: BehaviorSubject<boolean> = new BehaviorSubject(false);
+  get moduleImportIsAuthorized() {
+    return this._moduleImportIsAuthorized.getValue();
+  }
   
   constructor(
     private _dfs: DataFormService,
     private translate: TranslateService,
     public dialog: MatDialog,
+    public moduleService: ModuleService,
     public metadataS: MetadataService,
     public metadataDataS: MetadataDataService,
     private _commonService: CommonService,
   ) { }
 
   ngOnInit() {
-
+    //vérification que l'utilisateur est autorisé à utiliser le module d'import
+    this.moduleService.moduleSub
+      .pipe(
+        map((modules: any[]): boolean => {
+          if (modules) {
+            for (let i = 0; i < modules.length; i++) {
+              //recherche du module d'import et test si l'utilisateur a des droits dessus
+              if (modules[i].module_code == 'IMPORT' && modules[i].cruved['C'] > 0) {
+                return true;
+              }
+            }
+          }
+          return false;
+        })
+      )
+      .subscribe((importIsAuthorized: boolean) => this._moduleImportIsAuthorized.next(importIsAuthorized));
   }
